fix load_modules flaky
